+++ tests/testthat/test-as_tibble.R
@@ -122,25 +122,25 @@ test_that("as_tibble() checks for `unique` names by default (#278)", {
   l1 <- list(1:10)
   expect_tibble_error(
     as_tibble(l1),
-    error_column_names_cannot_be_empty(1)
+    error_column_names_cannot_be_empty(1, repair_hint = TRUE)
   )
 
   l2 <- list(x = 1, 2)
   expect_tibble_error(
     as_tibble(l2),
-    error_column_names_cannot_be_empty(2)
+    error_column_names_cannot_be_empty(2, repair_hint = TRUE)
   )
 
   l3 <- list(x = 1, ... = 2)
   expect_tibble_error(
     as_tibble(l3),
-    error_column_names_cannot_be_dot_dot(2)
+    error_column_names_cannot_be_dot_dot(2, repair_hint = TRUE)
   )
 
   l4 <- list(x = 1, ..1 = 2)
   expect_tibble_error(
     as_tibble(l4),
-    error_column_names_cannot_be_dot_dot(2)
+    error_column_names_cannot_be_dot_dot(2, repair_hint = TRUE)
   )
 
   df <- list(a = 1, b = 2)
@@ -148,7 +148,7 @@ test_that("as_tibble() checks for `unique` names by default (#278)", {
   df <- new_tibble(df, nrow = 1)
   expect_tibble_error(
     as_tibble(df),
-    error_column_names_cannot_be_empty(1:2)
+    error_column_names_cannot_be_empty(1:2, repair_hint = TRUE)
   )
 })
 
@@ -263,12 +263,12 @@ test_that("as_tibble.poly() supports .name_repair", {
 })
 
 test_that("as_tibble.table() supports .name_repair", {
-  x <- table(a = c(1, 1, 1, 2, 2, 2), a = c(3, 4, 5, 3, 4, 5))
+  expect_snapshot(error = TRUE, {
+    as_tibble(table(a = c(1, 1, 1, 2, 2, 2), a = c(3, 4, 5, 3, 4, 5)))
+    as_tibble(table(c(1, 1, 1, 2, 2, 2), c(3, 4, 5, 3, 4, 5)))
+  })
 
-  expect_tibble_error(
-    as_tibble(x),
-    error_column_names_must_be_unique("a")
-  )
+  x <- table(a = c(1, 1, 1, 2, 2, 2), a = c(3, 4, 5, 3, 4, 5))
   expect_identical(
     names(as_tibble(x, .name_repair = "minimal")),
     c("a", "a", "n")
@@ -410,7 +410,7 @@ test_that("as_tibble_row() can convert named bare vectors to data frame", {
   )
   expect_tibble_error(
     as_tibble_row(setNames(nm = c(TRUE, FALSE, NA))),
-    error_column_names_cannot_be_empty(3)
+    error_column_names_cannot_be_empty(3, repair_hint = TRUE)
   )
 })
 
@@ -453,7 +453,7 @@ test_that("`validate` triggers deprecation message, but then works", {
       "deprecated",
       fixed = TRUE
     ),
-    error_column_names_cannot_be_empty(2)
+    error_column_names_cannot_be_empty(2, repair_hint = TRUE)
   )
 
   expect_warning(
@@ -480,7 +480,7 @@ test_that("`validate` triggers deprecation message, but then works", {
       "deprecated",
       fixed = TRUE
     ),
-    error_column_names_cannot_be_empty(2)
+    error_column_names_cannot_be_empty(2, repair_hint = TRUE)
   )
 })
 
@@ -488,7 +488,7 @@ test_that("`validate` always raises lifecycle warning.", {
   expect_deprecated(
     expect_tibble_error(
       as_tibble(list(a = 1, "hi"), validate = TRUE, .name_repair = "check_unique"),
-      error_column_names_cannot_be_empty(2)
+      error_column_names_cannot_be_empty(2, repair_hint = TRUE)
     )
   )
 
@@ -511,7 +511,7 @@ test_that("`validate` always raises lifecycle warning.", {
       as_tibble(df, validate = TRUE, .name_repair = "check_unique"),
       NA
     ),
-    error_column_names_cannot_be_empty(2)
+    error_column_names_cannot_be_empty(2, repair_hint = TRUE)
   )
 })
 
@@ -519,7 +519,7 @@ test_that("Inconsistent `validate` and `.name_repair` used together raise a warn
   expect_deprecated(
     expect_tibble_error(
       as_tibble(list(a = 1, "hi"), validate = FALSE, .name_repair = "check_unique"),
-      error_column_names_cannot_be_empty(2)
+      error_column_names_cannot_be_empty(2, repair_hint = TRUE)
     )
   )
 
@@ -540,7 +540,7 @@ test_that("Inconsistent `validate` and `.name_repair` used together raise a warn
   expect_deprecated(
     expect_tibble_error(
       as_tibble(df, validate = FALSE, .name_repair = "check_unique"),
-      error_column_names_cannot_be_empty(2)
+      error_column_names_cannot_be_empty(2, repair_hint = TRUE)
     )
   )
 })