+++ tests/testthat/test-zzz-name-repair.R
@@ -21,12 +21,12 @@ test_that("set_minimal_names() copes with NULL input names", {
 test_that("check_minimal() errors when names aren't minimal", {
   expect_legacy_error(
     check_minimal(NULL),
-    error_names_must_be_non_null(repair = FALSE),
+    error_names_must_be_non_null(repair_hint = FALSE),
     fixed = TRUE
   )
   expect_legacy_error(
     check_minimal(c("a", NA)),
-    error_column_must_be_named(2, repair = FALSE),
+    error_column_must_be_named(2, repair_hint = FALSE),
     fixed = TRUE
   )
 })
@@ -76,13 +76,13 @@ test_that("unique_names() strips positional suffixes, re-applies as needed", {
 test_that("check_unique() imposes check_minimal()", {
   expect_legacy_error(
     check_unique(NULL),
-    error_names_must_be_non_null(repair = FALSE),
+    error_names_must_be_non_null(repair_hint = FALSE),
     fixed = TRUE
   )
 
   expect_legacy_error(
     check_unique(c("x", NA)),
-    error_column_must_be_named(2, repair = FALSE),
+    error_column_must_be_named(2, repair_hint = FALSE),
     fixed = TRUE
   )
 })
@@ -90,22 +90,22 @@ test_that("check_unique() imposes check_minimal()", {
 test_that("check_unique() errors for empty or duplicated names", {
   expect_legacy_error(
     check_unique(c("x", "")),
-    error_column_must_be_named(2, repair = FALSE),
+    error_column_must_be_named(2, repair_hint = FALSE),
     fixed = TRUE
   )
   expect_legacy_error(
     check_unique(c("", "x", "")),
-    error_column_must_be_named(c(1, 3), repair = FALSE),
+    error_column_must_be_named(c(1, 3), repair_hint = FALSE),
     fixed = TRUE
   )
   expect_legacy_error(
     check_unique(c("x", "x", "y")),
-    error_column_names_must_be_unique("x", repair = FALSE),
+    error_column_names_must_be_unique("x", repair_hint = FALSE),
     fixed = TRUE
   )
   expect_legacy_error(
     check_unique(c("x", "y", "x", "y")),
-    error_column_names_must_be_unique(c("x", "y"), repair = FALSE),
+    error_column_names_must_be_unique(c("x", "y"), repair_hint = FALSE),
     fixed = TRUE
   )
 })